+++ Anarchy/Properties/Stable/ChangeLog.md
@@ -1,2 +1,3 @@
 ﻿## Patch V1.7.23
-*   Added bridge for other mods to register their tool systems.
+*  Added bridge for other mods to register their tool systems.
+*  Added support for copying Anarchy and Transform Lock components using Move It Alpha's Copy/Paste.